ml_engine/requirements.txt
# NewsGuard ML Engine — Python Dependencies
# Install with:  pip install -r requirements.txt

# ── Core ML ───────────────────────────────────────
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# ── Text & NLP ────────────────────────────────────
nltk>=3.8.0

# ── API Server ────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# ── Visualisation (for training metrics) ──────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Optional: Web scraping (existing app.py) ──────
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# ── Optional: Deep Learning (LSTM model) ──────────
# tensorflow>=2.13.0   # Uncomment if you want LSTM fusion
